Added max hits and did a refactor for lots of code

+++ Rutschig/Config/Config.cs
@@ -4,7 +4,7 @@
     {
         public string AppName { get; set; }
         public string AppDomain { get; set; }
-        public short ShortenedLength { get; set; }
+        public byte ShortenedLength { get; set; }
         public int MaxPinLength { get; set; }
         public int MaxExpirationDelta { get; set; }
     }
